Return a dictionary from DuckDuckGoSearchTool
# Why?
The description is currently:
Performs a duckduckgo web search based on your query (think a Google search) then returns the top search results as a list of dict elements.  Each result has keys 'title', 'href' and 'body'.

But the return value was a Markdown string. As a result the code later was failing trying to access the 'body'/'href' key of the response

See issue #86

# What ?

The return value is a list of dict elements. This is as siginificantly different from the return value of the previous version of the tool.

Fix #86